refactor(autocomplete): reconfigure file structure
